Split up manageSpace.kt to separate class files and packages.
