+++ android/space/src/main/kotlin/net/twisterrob/inventory/android/activity/space/ManageSpaceUiEffect.kt
@@ -1,0 +1,7 @@
+package net.twisterrob.inventory.android.activity.space
+
+internal sealed class ManageSpaceUiEffect {
+	data class ShowToast(
+		val message: CharSequence
+	) : ManageSpaceUiEffect()
+}
