+++ android/space/src/main/kotlin/net/twisterrob/inventory/android/activity/space/sizes/GetSizesUseCase.kt
@@ -1,0 +1,74 @@
+package net.twisterrob.inventory.android.activity.space.sizes
+
+import android.content.Context
+import android.database.DatabaseUtils
+import android.database.sqlite.SQLiteDatabase
+import dagger.hilt.android.qualifiers.ApplicationContext
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.withContext
+import net.twisterrob.android.utils.tools.DatabaseTools
+import net.twisterrob.android.utils.tools.IOTools
+import net.twisterrob.inventory.android.Constants.Pic.GlideSetup
+import net.twisterrob.inventory.android.content.Database
+import net.twisterrob.inventory.android.viewmodel.UseCase
+import org.slf4j.LoggerFactory
+import java.io.File
+import javax.inject.Inject
+
+class GetSizesUseCase @Inject constructor(
+	@ApplicationContext private val context: Context,
+) : UseCase<Unit, SizesDomain> {
+
+	override suspend fun execute(input: Unit): SizesDomain = withContext(Dispatchers.IO) {
+		delay(1000) // STOPSHIP remove this
+		val database = Database.get(context)
+		SizesDomain(
+			imageCache = safe {
+				dirSizes(
+					GlideSetup.getCacheDir(context)
+				)
+			},
+			database = safe {
+				dirSizes(
+					context.getDatabasePath(database.helper.databaseName)
+				)
+			},
+			freelist = safe {
+				database.readableDatabase.let { it.freelistCount * it.pageSize }
+			},
+			searchIndex = safe {
+				database.searchSize
+			},
+			allData = safe {
+				dirSizes(
+					File(context.applicationInfo.dataDir),
+					context.externalCacheDir,
+					context.getExternalFilesDir(null),
+				)
+			},
+		)
+	}
+
+	private fun dirSizes(vararg dirs: File?): Long =
+		dirs.sumOf { IOTools.calculateSize(it) }
+
+	private val SQLiteDatabase.freelistCount: Long
+		get() = DatabaseUtils.longForQuery(
+			this,
+			"PRAGMA freelist_count;",
+			DatabaseTools.NO_ARGS
+		)
+
+	private fun safe(block: () -> Long): Result<Long> =
+		try {
+			Result.success(block())
+		} catch (@Suppress("TooGenericExceptionCaught") ex: Exception) {
+			LOG.error("Cannot get size", ex)
+			Result.failure(ex)
+		}
+
+	companion object {
+		private val LOG = LoggerFactory.getLogger(GetSizesUseCase::class.java)
+	}
+}
