+++ android/space/src/main/kotlin/net/twisterrob/inventory/android/activity/space/sizes/SizesDomain.kt
@@ -1,0 +1,9 @@
+package net.twisterrob.inventory.android.activity.space.sizes
+
+data class SizesDomain(
+	val imageCache: Result<Long>,
+	val database: Result<Long>,
+	val freelist: Result<Long>,
+	val searchIndex: Result<Long>,
+	val allData: Result<Long>,
+)
